+++ modules/siddhi-core/src/test/java/io/siddhi/core/query/table/teststorecontaininginmemorytable/JoinTestStoreTestCase.java
@@ -62,8 +62,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -128,8 +126,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -193,8 +189,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol1 string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -257,8 +251,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -320,8 +312,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -383,8 +373,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -446,8 +434,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -511,8 +497,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -574,8 +558,6 @@ public class JoinTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "@connection(maxWait = '4000')" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
