+++ modules/siddhi-core/src/test/java/io/siddhi/core/query/table/teststorecontaininginmemorytable/UpdateTestStoreTestCase.java
@@ -63,8 +63,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price double, volume long); " +
                 "define stream UpdateStockStream (symbol string, price double, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price double, volume long); ";
 
         String query = "" +
@@ -104,8 +102,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price double, volume long); " +
                 "define stream UpdateStockStream (symbol string, price double, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price double, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -144,8 +140,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -196,8 +190,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -249,8 +241,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -301,8 +291,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream CheckStockStream (symbol string, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -355,8 +343,6 @@ public class UpdateTestStoreTestCase {
                 "define stream CheckStockStream (symbol string, price float, volume long); " +
                 "define stream UpdateStockStream (comp string, prc float, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -431,8 +417,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price float, volume long); " +
                 "define stream UpdateStockStream (symbol string, price float, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price float, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
@@ -471,8 +455,6 @@ public class UpdateTestStoreTestCase {
                 "define stream StockStream (symbol string, price double, volume long); " +
                 "define stream UpdateStockStream (symbol string, price double, volume long); " +
                 "@Store(type=\"testStoreContainingInMemoryTable\")\n" +
-                //"@PrimaryKey(\"symbol\")" +
-                //"@Index(\"volume\")" +
                 "define table StockTable (symbol string, price double, volume long); ";
         String query = "" +
                 "@info(name = 'query1') " +
